+++ components/QuickWin/index.js
@@ -1,0 +1,1 @@
+export { QuickWin } from './QuickWin';